Move DER encoding/parsing functions to BinaryUtil
